% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_lipid_mass.R
\name{calc_lipid_masses}
\alias{calc_lipid_masses}
\title{Calculation of mass of intact lipids

This function calcluates the mass of an intact lipid. Supported modifications are currently hydroxy groups (OH), hydroperoxy groups (OOH), keto groups (O) and amino groups (NH2).}
\usage{
calc_lipid_masses(lipids)
}
\arguments{
\item{lipid}{Lipid for which the mass shall be calculated.}
}
\description{
Calculation of mass of intact lipids

This function calcluates the mass of an intact lipid. Supported modifications are currently hydroxy groups (OH), hydroperoxy groups (OOH), keto groups (O) and amino groups (NH2).
}
\examples{
library(lipidomicsUtils)
calc_lipid_mass("CoA(18:1(9Z))")
calc_lipid_mass("PC(16:0/18:1(9Z))")

}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
