% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_acyl_mass.R
\name{calc_residue_acyl_mass}
\alias{calc_residue_acyl_mass}
\title{Calculation of acyl, alkyl or alkenyl mass (residue fatty acid or fatty alcohol)

This function calculates the mass of an acyl, alkyl or alkenyl given as shorthand notation, e.g. "18:1(9Z)" and calculates the mass of the respective residue fatty acid or alcohol. Supported modifications are currently hydroxy groups (OH), hydroperoxy groups (OOH), keto groups (O) and amino groups (NH2).}
\usage{
calc_residue_acyl_mass(x)
}
\arguments{
\item{x}{Shorthand notation of a acyl, alkyl, alkenyl (as string), e.g. "18:1(9Z)"}
}
\description{
Calculation of acyl, alkyl or alkenyl mass (residue fatty acid or fatty alcohol)

This function calculates the mass of an acyl, alkyl or alkenyl given as shorthand notation, e.g. "18:1(9Z)" and calculates the mass of the respective residue fatty acid or alcohol. Supported modifications are currently hydroxy groups (OH), hydroperoxy groups (OOH), keto groups (O) and amino groups (NH2).
}
\examples{
library(lipidomicsUtils)
calc_residue_acyl_mass("18:1(9Z)")

}
\seealso{
\code{\link{calc_intact_acyl_mass}}

\code{\link{calc_intact_acyl_formula}}

\code{\link{calc_residue_acyl_formula}}
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
