% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_sphingoid_mass.R
\name{calc_sphingoid_mass}
\alias{calc_sphingoid_mass}
\title{Calculation of sphingoid mass

This function calculates the mass of a sphingoid base.}
\usage{
calc_sphingoid_mass(sphingoid_base)
}
\arguments{
\item{lipid}{Lipid for which the mass shall be calculated.}
}
\description{
Calculation of sphingoid mass

This function calculates the mass of a sphingoid base.
}
\examples{
library(lipidomicsUtils)
calc_sphingoid_mass("d16:1(1OH,3OH,15Me)")

}
\seealso{
\code{\link{calc_sphingoid_formula}}
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
