% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_acyl_mass.R
\name{.acyl_formula}
\alias{.acyl_formula}
\title{Private function for calculation of intact acyl formula}
\usage{
.acyl_formula(x)
}
\description{
Private function for calculation of intact acyl formula
}
