% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_acyl_mass.R
\name{.alkenyl_formula}
\alias{.alkenyl_formula}
\title{Private function for calculation of intact alkenyl formula}
\usage{
.alkenyl_formula(x)
}
\description{
Private function for calculation of intact alkenyl formula
}
