% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_acyl_mass.R
\name{.alkenyl_mass}
\alias{.alkenyl_mass}
\title{Private function for calculation of intact alkenyl mass}
\usage{
.alkenyl_mass(x)
}
\description{
Private function for calculation of intact alkenyl mass
}
