% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_acyl_mass.R
\name{.alkyl_formula}
\alias{.alkyl_formula}
\title{Private function for calculation of intact alkyl formula}
\usage{
.alkyl_formula(x)
}
\description{
Private function for calculation of intact alkyl formula
}
