% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shorthand_fa.R
\name{.get_fa_species_level}
\alias{.get_fa_species_level}
\title{Private function to generate species level shorthand}
\usage{
.get_fa_species_level(x)
}
\description{
Private function to generate species level shorthand
}
