% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parser_functional_groups.R
\name{get_amino_number}
\alias{get_amino_number}
\title{Get number of keto groups

This function returns the number of keto groups in a acyl, alkyl, alkenyl, sphingoid base etc.}
\usage{
get_amino_number(x)
}
\arguments{
\item{lipid}{Lipid for which the mass shall be calculated.}
}
\description{
Get number of keto groups

This function returns the number of keto groups in a acyl, alkyl, alkenyl, sphingoid base etc.
}
\examples{
library(lipidomicsUtils)
get_amino_number("18:0(2NH2)")

}
\seealso{
\code{\link{get_carbon_number}}

\code{\link{get_bond_number}}

\code{\link{get_hydroxy_number}}

\code{\link{get_peroxy_number}}

\code{\link{get_keto_number}}
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
