% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shorthand_fa.R
\name{get_fa_biggid}
\alias{get_fa_biggid}
\title{Generation of a BiGG like ID for an acyl based shorthand notation

This function generates a BiGG like ID for GSMNs for a given shorthand notation, e.g. CoA(18:2(9Z,12Z)) to coa18_2_9z12z. Supported modifications are currently hydroxy groups (OH), hydroperoxy groups (OOH), keto groups (O) and amino groups (NH2).}
\usage{
get_fa_biggid(x)
}
\arguments{
\item{x}{Shorthand notation of a fatty acyl lipid, e.g. CoA(18:2(9Z,12Z))}
}
\description{
Generation of a BiGG like ID for an acyl based shorthand notation

This function generates a BiGG like ID for GSMNs for a given shorthand notation, e.g. CoA(18:2(9Z,12Z)) to coa18_2_9z12z. Supported modifications are currently hydroxy groups (OH), hydroperoxy groups (OOH), keto groups (O) and amino groups (NH2).
}
\examples{
library(lipidomicsUtils)
get_fa_biggid("CoA(18:2(9Z,12Z))")

}
\seealso{
\code{\link{get_fa_shorthand}}
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
