% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shorthand_fa.R
\name{get_fa_shorthand}
\alias{get_fa_shorthand}
\title{Generation of different shorthand notation level

This function simplifies a given shorthand notation to a lower level notation, e.g. CoA(18:2(9Z,12Z)) to CoA(18:2). Supported modifications are currently hydroxy groups (OH), hydroperoxy groups (OOH), keto groups (O) and amino groups (NH2).}
\usage{
get_fa_shorthand(x, level = c("structural", "molecular", "species"))
}
\arguments{
\item{x}{Shorthand notation of a fatty acyl lipid, e.g. CoA(18:2(9Z,12Z))}

\item{level}{Either \code{structural}, \code{molecular} or \code{species} for the respective levels according to SwissLipids hierachy}
}
\description{
Generation of different shorthand notation level

This function simplifies a given shorthand notation to a lower level notation, e.g. CoA(18:2(9Z,12Z)) to CoA(18:2). Supported modifications are currently hydroxy groups (OH), hydroperoxy groups (OOH), keto groups (O) and amino groups (NH2).
}
\examples{
library(lipidomicsUtils)
get_fa_shorthand("CoA(18:2(9Z,12Z))")

}
\seealso{
\code{\link{get_fa_biggid}}
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
