% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shorthand_gp.R
\name{get_gp_shorthand}
\alias{get_gp_shorthand}
\title{Generation of different shorthand notation level

This function simplifies a given shorthand notation to a lower level notation, e.g. PC(16:0/18:2(9Z,12Z)) to PC(16:0/18:2) or PC(16:0_18:2) or PC(34:2).}
\usage{
get_gp_shorthand(x, level = c("structural", "molecular", "species"))
}
\arguments{
\item{x}{Shorthand notation of a fatty acyl lipid, e.g. PC(16:0/18:2(9Z,12Z))}

\item{level}{Either \code{structural}, \code{molecular} or \code{species} for the respective levels according to SwissLipids hierachy}
}
\description{
Generation of different shorthand notation level

This function simplifies a given shorthand notation to a lower level notation, e.g. PC(16:0/18:2(9Z,12Z)) to PC(16:0/18:2) or PC(16:0_18:2) or PC(34:2).
}
\examples{
library(lipidomicsUtils)
get_gl_shorthand("PC(16:0/18:2(9Z,12Z))")

}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
