% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parser_functional_groups.R
\name{get_keto_groups}
\alias{get_keto_groups}
\title{Get all keto groups

This function returns all keto groups in a acyl, alkyl, alkenyl, sphingoid base etc.}
\usage{
get_keto_groups(x)
}
\arguments{
\item{lipid}{Lipid for which the mass shall be calculated.}
}
\description{
Get all keto groups

This function returns all keto groups in a acyl, alkyl, alkenyl, sphingoid base etc.
}
\examples{
library(lipidomicsUtils)
get_keto_groups("18:0(3O)")

}
\seealso{
\code{\link{get_methyl_branches}}

\code{\link{get_double_bonds}}

\code{\link{get_hydroxy_groups}}

\code{\link{get_peroxy_groups}}

\code{\link{get_amino_groups}}
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
