% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_lipid_category.R
\name{get_lipid_category}
\alias{get_lipid_category}
\title{Determine the lipid category

This function determines the lipid category to which a lipid given as shorthand notation belongs}
\usage{
get_lipid_category(lipid)
}
\arguments{
\item{lipid}{Lipid for which the category shall be determined.}
}
\description{
Determine the lipid category

This function determines the lipid category to which a lipid given as shorthand notation belongs
}
\examples{
library(lipidomicsUtils)
get_lipid_category("PC(16:0/18:1(9Z)")

}
\seealso{
\code{\link{get_lipid_mainclass}}

\code{\link{get_lipid_subclass}}
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
