% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_lipid_category.R
\name{get_lipid_subclass}
\alias{get_lipid_subclass}
\title{Determine the lipid sub class

This function determines the lipid sub class to which a lipid given as shorthand notation belongs}
\usage{
get_lipid_subclass(lipid)
}
\arguments{
\item{lipid}{Lipid for which the sub class shall be determined.}
}
\description{
Determine the lipid sub class

This function determines the lipid sub class to which a lipid given as shorthand notation belongs
}
\examples{
library(lipidomicsUtils)
get_lipid_category("PC(16:0/18:1(9Z)")

}
\seealso{
\code{\link{get_lipid_category}}

\code{\link{get_lipid_mainclass}}
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
