% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parser_functional_groups.R
\name{get_methyl_branches}
\alias{get_methyl_branches}
\title{Get all methyl branches

This function returns all methyl branches in a acyl, alkyl, alkenyl, sphingoid base etc.}
\usage{
get_methyl_branches(x)
}
\arguments{
\item{lipid}{Lipid for which the mass shall be calculated.}
}
\description{
Get all methyl branches

This function returns all methyl branches in a acyl, alkyl, alkenyl, sphingoid base etc.
}
\examples{
library(lipidomicsUtils)
get_methyl_branches("d16:1(1OH,3OH,15Me)")

}
\seealso{
\code{\link{get_double_bonds}}

\code{\link{get_hydroxy_groups}}

\code{\link{get_keto_groups}}

\code{\link{get_peroxy_groups}}

\code{\link{get_amino_groups}}
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
