% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parser_functional_groups.R
\name{get_peroxy_groups}
\alias{get_peroxy_groups}
\title{Get all peroxy groups

This function returns all peroxy groups in a acyl, alkyl, alkenyl, sphingoid base etc.}
\usage{
get_peroxy_groups(x)
}
\arguments{
\item{lipid}{Lipid for which the mass shall be calculated.}
}
\description{
Get all peroxy groups

This function returns all peroxy groups in a acyl, alkyl, alkenyl, sphingoid base etc.
}
\examples{
library(lipidomicsUtils)
get_peroxy_groups("18:0(5OOH)")

}
\seealso{
\code{\link{get_methyl_branches}}

\code{\link{get_double_bonds}}

\code{\link{get_hydroxy_groups}}

\code{\link{get_keto_groups}}

\code{\link{get_amino_groups}}
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
