% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shorthand_sph.R
\name{get_sph_shorthand}
\alias{get_sph_shorthand}
\title{Generation of different shorthand notation level

This function simplifies a given shorthand notation to a lower level notation, e.g. Cer(d16:1(4E,1OH,3OH,15Me)/22:0) to Cer(d17:1/22:0) or Cer(d39:1).}
\usage{
get_sph_shorthand(x, level = c("structural", "molecular", "species"))
}
\arguments{
\item{x}{Shorthand notation of a fatty acyl lipid, e.g. Cer(d16:1(4E,1OH,3OH,15Me)/22:0)}

\item{level}{Either \code{structural}, \code{molecular} or \code{species} for the respective levels according to SwissLipids hierachy}
}
\description{
Generation of different shorthand notation level

This function simplifies a given shorthand notation to a lower level notation, e.g. Cer(d16:1(4E,1OH,3OH,15Me)/22:0) to Cer(d17:1/22:0) or Cer(d39:1).
}
\examples{
library(lipidomicsUtils)
get_sph_shorthand("Cer(d16:1(4E,1OH,3OH,15Me)/22:0)")

}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
