% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_swiss_lipids.R
\name{inchikey_to_swissid_exact}
\alias{inchikey_to_swissid_exact}
\title{Search for SwissLipids ID from InChIKey (exact)

This function uses the SwissLipids API to search for a SwissLipids ID from a given InChIKey. An exact match of the InChIKey is expected.}
\usage{
inchikey_to_swissid_exact(inchikey)
}
\arguments{
\item{inchikey}{InChIKey for which a SwissLipids ID shall be searched for}
}
\description{
Search for SwissLipids ID from InChIKey (exact)

This function uses the SwissLipids API to search for a SwissLipids ID from a given InChIKey. An exact match of the InChIKey is expected.
}
\examples{
library(lipidomicsUtils)
inchikey_to_swissid_exact("WTJKGGKOPKCXLL-VYOBOKEXSA-N")

}
\seealso{
\code{\link{inchikey_to_swissid_nocharge}}

\code{\link{inchikey_to_swissid_structure}}
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
