% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_swiss_lipids.R
\name{inchikey_to_swissid_structure}
\alias{inchikey_to_swissid_structure}
\title{Search for SwissLipids ID from InChIKey (only atom connectivity)

This function uses the SwissLipids API to search for a SwissLipids ID from a given InChIKey. Only the atom connectivity block is used.}
\usage{
inchikey_to_swissid_structure(inchikey)
}
\arguments{
\item{inchikey}{InChIKey for which a SwissLipids ID shall be searched for}
}
\description{
Search for SwissLipids ID from InChIKey (only atom connectivity)

This function uses the SwissLipids API to search for a SwissLipids ID from a given InChIKey. Only the atom connectivity block is used.
}
\examples{
library(lipidomicsUtils)
inchikey_to_swissid_structure("WTJKGGKOPKCXLL-VYOBOKEXSA-N")

 @author Michael Witting, \\email{michael.witting@helmholtz-muenchen.de}

}
\seealso{
\code{\link{inchikey_to_swissid_exact}}

\code{\link{inchikey_to_swissid_nocharge}}
}
