% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parser_fatty_acyl.R
\name{isolate_radyls}
\alias{isolate_radyls}
\title{Get all fatty acyls, alkyls and alkenyls

This functions isolates all fatty acyls, alkyls and alkenyls from a given lipid shorthand notation and returns them as vector. upported modifications are currently hydroxy groups (OH), hydroperoxy groups (OOH), keto groups (O) and amino groups (NH2)}
\usage{
isolate_radyls(lipids)
}
\arguments{
\item{lipids}{List of vector of shorthand notation of a acyl (as string), e.g. "c(PC(18:0/20:4(7E,9E,11Z,14Z)), "PC(16:0/18:1(9Z))")}
}
\description{
Get all fatty acyls, alkyls and alkenyls

This functions isolates all fatty acyls, alkyls and alkenyls from a given lipid shorthand notation and returns them as vector. upported modifications are currently hydroxy groups (OH), hydroperoxy groups (OOH), keto groups (O) and amino groups (NH2)
}
\examples{
library(lipidomicsUtils) 
lipids <- c(PC(18:0/20:4(7E,9E,11Z,14Z)), "PC(16:0/18:1(9Z))")
isolate_radyls(lipids)

}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
