% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parser_sphingoid_base.R
\name{isolate_sphingoids}
\alias{isolate_sphingoids}
\title{Get all sphingoid bases

This functions isolates all sphingoid bases from a given lipid shorthand notation and returns them as vector. upported modifications are currently hydroxy groups (OH), hydroperoxy groups (OOH), keto groups (O) and amino groups (NH2)}
\usage{
isolate_sphingoids(lipids)
}
\arguments{
\item{lipid}{Shorthand notation of a acyl (as string), e.g. "Cer(d16:1(4E,1OH,3OH,15Me)/22:0)"}
}
\description{
Get all sphingoid bases

This functions isolates all sphingoid bases from a given lipid shorthand notation and returns them as vector. upported modifications are currently hydroxy groups (OH), hydroperoxy groups (OOH), keto groups (O) and amino groups (NH2)
}
\examples{
library(lipidomicsUtils) 
isolate_fatty_acyls("Cer(d16:1(4E,1OH,3OH,15Me)/22:0)")

}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
