% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_sphingoid_mass.R
\name{calc_sphingoid_mass}
\alias{calc_sphingoid_mass}
\title{Calculation of sphingoid base mass (intact sphingoid base)

This function calculates the mass of a sphingnoid base given as shorthand notation, e.g. "d16:1(4E)(15Me)" and calculates the mass of the respective sphingoid base. Supported modifications are currently hydroxy groups (OH), hydroperoxy groups (OOH), keto groups (O) and amino groups (NH2).}
\usage{
calc_sphingoid_mass(x)
}
\arguments{
\item{x}{Shorthand notation of a sphingoid base (as string), e.g. "d16:1(4E)(15Me)"}
}
\description{
Calculation of sphingoid base mass (intact sphingoid base)

This function calculates the mass of a sphingnoid base given as shorthand notation, e.g. "d16:1(4E)(15Me)" and calculates the mass of the respective sphingoid base. Supported modifications are currently hydroxy groups (OH), hydroperoxy groups (OOH), keto groups (O) and amino groups (NH2).
}
