% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parser_functional_groups.R
\name{get_carbon_number}
\alias{get_carbon_number}
\title{Get number of carbons

This function returns the number of carbons in a acyl, alkyl, alkenyl or sphingoid base based on a supplied shorthand notation, e.g. d16:1(4E)(1OH,3OH)(15Me)}
\usage{
get_carbon_number(x)
}
\arguments{
\item{x}{Shorthand notation of a acyl, alkyl, alkenyl (as string), e.g. "d16:1(4E)(1OH,3OH)(15Me)"}
}
\description{
Get number of carbons

This function returns the number of carbons in a acyl, alkyl, alkenyl or sphingoid base based on a supplied shorthand notation, e.g. d16:1(4E)(1OH,3OH)(15Me)
}
